pyproject.toml
[project]
name = "axatechtask"
version = "0.1.0"
description = ""
authors = [
    {name = "saddiqs1",email = "ssaddique96@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false